Move ViewDefinition file
